Add a Sentinel policy to protect the policy management workspace
Since the config that manages Sentinel policies has to have a `tfe` provider
authenticated with a highly privileged token (org token, owners team token, or
owner user token), it could be used for a variety of evil or just accidentally
bad purposes. (Privilege escalation by managing tfe_team_access resources, or
really anything.)

So we'll add a Sentinel policy to restrict what that workspace can be used for.