Revise all Sentinel policies for deep search in prod and easy mocks in test
Sing it along with me again: you have to iterate through the modules if you
really want to test ALL instances of a resource type, but since you can't mock
functions in a JSON config file, doing the right thing makes your policy
untestable.

But there's a way around that, which is to check whether we've loaded the real
import and fall back to just checking `tfplan.resources` if not. This gives us a
real global resource search in TFE, while continuing to work with the existing
mock data. (Note that our tests all pass.)

I'm using `tfplan.module_paths` as a proxy for the real import. This makes an
assumption that you're not mocking `module_paths` in your test files, but that
seems like a reasonable trade-off for policies that are mainly about global
policy on resources.