Refactor out the nested `all` expressions in policies
In all of these cases, we were having to deal with deeply nested data structures
way too late in the policy, making the _policy_ part of it harder to read. By
moving all the nested data structure logic into one helper function, we can make
the important part of the policy more legible.